requirements.txt
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
pandas==2.1.3
schedule==1.2.0
flask==3.0.0
flask-cors==4.0.0

